% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{ad}
\alias{ad}
\title{Accumulation/Distribution Line}
\usage{
ad(high, low, close, volume)
}
\arguments{
\item{high}{high}

\item{low}{low}

\item{close}{close}

\item{volume}{volume}
}
\value{
numeric vector
}
\description{
Accumulation/Distribution Line
}
