% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{adosc}
\alias{adosc}
\title{Accumulation/Distribution Oscillator}
\usage{
adosc(high, low, close, volume, period_short, period_long)
}
\arguments{
\item{high}{high}

\item{low}{low}

\item{close}{close}

\item{volume}{volume}

\item{period_short}{short smoothing period}

\item{period_long}{long smoothing period}
}
\value{
numeric vector
}
\description{
Accumulation/Distribution Oscillator
}
