% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{ao}
\alias{ao}
\title{Awesome Oscillator}
\usage{
ao(high, low, short_period = 5L, long_period = 34L)
}
\arguments{
\item{high}{high}

\item{low}{low}

\item{short_period}{short smoothing period}

\item{long_period}{long smoothing period}
}
\value{
numeric vector
}
\description{
Awesome Oscillator
}
