% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{arosc}
\alias{arosc}
\title{Aroon Oscillator}
\usage{
arosc(high, low, period)
}
\arguments{
\item{high}{high}

\item{low}{low}

\item{period}{period}
}
\value{
numeric vector
}
\description{
Aroon Oscillator
}
