% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{ashare_intraday_tsec}
\alias{ashare_intraday_tsec}
\title{Convert A-share intraday timestamps to seconds}
\usage{
ashare_intraday_tsec(t)
}
\arguments{
\item{t}{a vector of timestamps}
}
\value{
a vector of integer, counted seconds of t
}
\description{
9:30 is converted to 0th second, 13:00 is converted to 7201st second
}
