% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{atr}
\alias{atr}
\title{Average True Range}
\usage{
atr(high, low, close, period)
}
\arguments{
\item{high}{high}

\item{low}{low}

\item{close}{close}

\item{period}{period}
}
\value{
numeric vector
}
\description{
Average True Range
}
