% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tushare_api.R
\name{$.tsapi}
\alias{$.tsapi}
\title{Request data from Tushare API}
\usage{
\method{$}{tsapi}(x, func)
}
\arguments{
\item{x}{A tsapi object}

\item{func}{Tushare API function to call}
}
\value{
a data.table
}
\description{
Request data from Tushare API
}
