% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{cci}
\alias{cci}
\title{Commodity Channel Index}
\usage{
cci(x, period, mdr = 0.015)
}
\arguments{
\item{x}{measured variable / typical price}

\item{period}{period}

\item{mdr}{a ratio to adjust ±100 region}
}
\value{
numeric vector
}
\description{
Commodity Channel Index
}
