% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{cmo}
\alias{cmo}
\title{Chande Momentum Oscillator}
\usage{
cmo(x, period)
}
\arguments{
\item{x}{measured variable}

\item{period}{period}
}
\value{
numeric vector
}
\description{
Chande Momentum Oscillator
}
