% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_loop.R
\name{connect_rt_db}
\alias{connect_rt_db}
\title{Return DBI connection to realtime database}
\usage{
connect_rt_db(db = get_rt_db())
}
\arguments{
\item{db}{path to database}
}
\value{
a DBI connection
}
\description{
Return DBI connection to realtime database
}
