% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_sina.R
\name{connect_srt_db}
\alias{connect_srt_db}
\title{Connect to Sina realtime database}
\usage{
connect_srt_db(db = get_srt_db())
}
\arguments{
\item{db}{file path to database}
}
\value{
a DBI connection
}
\description{
Connect to Sina realtime database
}
