% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{create_index}
\alias{create_index}
\title{Create index on a table}
\usage{
create_index(con, name, tbl, var, unique = FALSE)
}
\arguments{
\item{con}{an SQL connection}

\item{name}{name of index}

\item{tbl}{name of table}

\item{var}{variables/columns to create index}

\item{unique}{whether to create a unique index}
}
\value{
TRUE/FALSE
}
\description{
Create index on a table
}
