% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_loop.R
\name{create_rt_loader}
\alias{create_rt_loader}
\title{Create a realtime data loader}
\usage{
create_rt_loader(
  db = get_rt_db(),
  tbl_name = "rtdt",
  code_col = "Code",
  time_col = "Time",
  today = idate_now(),
  api = TushareApi()
)
}
\arguments{
\item{db}{path to database}

\item{tbl_name}{table name that stores realtime data}

\item{code_col}{column name to code}

\item{time_col}{column name to time}

\item{today}{date of today}

\item{api}{a tsapi object}
}
\value{
a data loader function, which returns incremental data.table of new data
}
\description{
Create a realtime data loader
}
