% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_sina.R
\name{create_srt_db}
\alias{create_srt_db}
\title{Create a Sina realtime database}
\usage{
create_srt_db(db = get_srt_db(), api = TushareApi())
}
\arguments{
\item{db}{file path to database}

\item{api}{a tsapi object}
}
\value{
TRUE
}
\description{
Create a Sina realtime database
}
