% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{csv_data_writer}
\alias{csv_data_writer}
\title{Generate a simple csv writer for sina_realtime_worker()}
\usage{
csv_data_writer(file = tempfile(fileext = "csv"))
}
\arguments{
\item{file}{output csv file path}
}
\value{
a data_write function that appends Sina quote data to csv file
}
\description{
Generate a simple csv writer for sina_realtime_worker()
}
\examples{
data_writer <- csv_data_write("records.csv")
}
