% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_algo.R
\name{cumulative_moment}
\alias{cumulative_moment}
\title{Cumulative moments for up to 4th order}
\usage{
cumulative_moment(order = 4L)
}
\arguments{
\item{order}{order of moments}
}
\value{
an online function
}
\description{
Cumulative moments for up to 4th order
}
