% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eod.R
\name{daily}
\alias{daily}
\title{Query daily OHLC data}
\usage{
daily(
  ts_code,
  start_date = "",
  end_date = "",
  adjust = c("none", "forward", "qfq", "backward", "hfq"),
  api = TushareApi()
)
}
\arguments{
\item{ts_code}{Tushare code}

\item{start_date}{OPTIONAL, start date of OHLC}

\item{end_date}{OPTIONAL, end date of OHLC}

\item{adjust}{Dividen adjustment method}

\item{api}{a tsapi object}
}
\value{
a data.table
}
\description{
Query daily OHLC data
}
