% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapdv.R
\name{dapdv_plot}
\alias{dapdv_plot}
\title{Plot dapdv prob density distribution}
\usage{
dapdv_plot(dt, plot = c("ROE", "PB"))
}
\arguments{
\item{dt}{result from dapdv}

\item{plot}{value to plot}
}
\value{
a ggplot object
}
\description{
Plot dapdv prob density distribution
}
