% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_sina.R
\name{default_srt_codes}
\alias{default_srt_codes}
\title{Query default Sina realtime codes}
\usage{
default_srt_codes(stock = TRUE, fund = TRUE, index = TRUE)
}
\arguments{
\item{stock}{whether to include stocks}

\item{fund}{whether to include funds}

\item{index}{whether to include indices}
}
\value{
a character vector
}
\description{
Query default Sina realtime codes
}
