% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{dpo}
\alias{dpo}
\title{Detrended Price Oscillator}
\usage{
dpo(x, period)
}
\arguments{
\item{x}{measured variable}

\item{period}{period}
}
\value{
numeric vector
}
\description{
Detrended Price Oscillator
}
