% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{emv}
\alias{emv}
\title{Ease of Movement}
\usage{
emv(high, low, volume, lot = 100, lag = 1L)
}
\arguments{
\item{high}{high}

\item{low}{low}

\item{volume}{volume}

\item{lot}{volume per lot}

\item{lag}{lag window}
}
\value{
numeric vector
}
\description{
Ease of Movement
}
