% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{example_data_handler}
\alias{example_data_handler}
\title{An example data_handler for sina_realtime_worker()}
\usage{
example_data_handler(frame)
}
\arguments{
\item{frame}{a frame of realtime quote data passed by worker}
}
\value{
a named list, which is further converted to a data.table then passed to result_reporter
}
\description{
An example data_handler for sina_realtime_worker()
}
