% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{example_result_reporter}
\alias{example_result_reporter}
\title{An example result_reporter for sina_realtime_worker()}
\usage{
example_result_reporter(result)
}
\arguments{
\item{result}{result passed by worker}
}
\description{
An example result_reporter for sina_realtime_worker()
}
