% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{example_signal_handler}
\alias{example_signal_handler}
\title{An example signal_handler for sina_realtime_worker()}
\usage{
example_signal_handler(result)
}
\arguments{
\item{result}{result passed by worker}
}
\description{
An example signal_handler for sina_realtime_worker()
}
