% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdepth.R
\name{filter_bookdepth}
\alias{filter_bookdepth}
\title{Filter order book / tick data by available book depth}
\usage{
filter_bookdepth(dt, avail_depth = book_depth, book_depth = 5L)
}
\arguments{
\item{dt}{order book or tick data}

\item{avail_depth}{minimum available book depth}

\item{book_depth}{max book depth}
}
\value{
a data.table
}
\description{
Filter order book / tick data by available book depth
}
