% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_loop.R
\name{get_rt_db}
\alias{get_rt_db}
\title{Return realtime database}
\usage{
get_rt_db(wd = getwd(), filename = paste0(Sys.Date(), ".db"))
}
\arguments{
\item{wd}{working directory}

\item{filename}{filename of database}
}
\value{
a character vector
}
\description{
Return realtime database
}
