% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{get_sina_code}
\alias{get_sina_code}
\title{Convert symbols to Sina codes}
\usage{
get_sina_code(symbol)
}
\arguments{
\item{symbol}{a vector of character symbols}
}
\value{
a vector of Sina codes
}
\description{
Convert symbols to Sina codes
}
