% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_sina.R
\name{get_srt_db}
\alias{get_srt_db}
\title{Return default Sina realtime database file path}
\usage{
get_srt_db(wd = getwd(), filename = paste0(Sys.Date(), ".db"))
}
\arguments{
\item{wd}{a working directory}

\item{filename}{database filename}
}
\value{
a string, file path to database
}
\description{
Return default Sina realtime database file path
}
