% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{group_by_interval}
\alias{group_by_interval}
\title{Group time by time intervals}
\usage{
group_by_interval(t, intv, grp_id = TRUE)
}
\arguments{
\item{t}{a vector of POSIXct/ITime}

\item{intv}{time intervals to group}

\item{grp_id}{wheter to return group id or grouped time}
}
\value{
a vector of group id or POSIXct/ITime
}
\description{
Group time by time intervals
}
