% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{grp_interval}
\alias{grp_interval}
\title{Group variable by intervals, left open}
\usage{
grp_interval(x, intv, grp_id = TRUE)
}
\arguments{
\item{intv}{time intervals to group}

\item{grp_id}{wheter to return group id or grouped value}

\item{t}{a vector of POSIXct/ITime}
}
\value{
a vector of group id or POSIXct/ITime
}
\description{
Group variable by intervals, left open
}
