% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{insert_to}
\alias{insert_to}
\title{Insert data to an SQL connection}
\usage{
insert_to(
  con,
  tbl,
  dt,
  conflict = c("default", "replace", "ignore"),
  constraint = data.table::key(dt)
)
}
\arguments{
\item{con}{an SQL connection}

\item{tbl}{name of table to insert to}

\item{dt}{data to insert}

\item{conflict}{how to deal with conflict}

\item{constraint}{constrained columns, only used when conflict == "replace" and connection is to a PostgreSQL database}
}
\value{
number of rows inserted
}
\description{
Insert data to an SQL connection
}
