% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eod.R
\name{intraday}
\alias{intraday}
\title{Query intraday OHLC data}
\usage{
intraday(
  api,
  ts_code,
  start_date = "",
  end_date = "",
  freq = c("1", "5", "15", "60")
)
}
\arguments{
\item{api}{a tsapi object}

\item{ts_code}{Tushare code}

\item{start_date}{OPTIONAL, start time of OHLC}

\item{end_date}{OPTIONAL, end time of OHLC}

\item{freq}{frequency of OHLC}
}
\value{
a data.table
}
\description{
Query intraday OHLC data
}
