% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_loop.R
\name{itime_now}
\alias{itime_now}
\alias{idate_now}
\title{Return current time/date as ITime/IDate}
\usage{
itime_now(api = TushareApi())

idate_now()
}
\arguments{
\item{api}{a tsapi object}
}
\value{
ITime/IDate
}
\description{
Return current time/date as ITime/IDate
}
