% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_lag.R
\name{lag_delta}
\alias{lag_delta}
\alias{lag_ratio}
\alias{lag_delta_moving_sum}
\title{Calculate lagged element/delta/ratio}
\usage{
lag_delta(x, lag, na_fill = NA)

lag_ratio(x, lag, na_fill = NA)

lag_delta_moving_sum(x, window, lag)
}
\arguments{
\item{x}{measured variable}

\item{lag}{lag period}

\item{na_fill}{filled value when lagged data is not available}

\item{window}{moving window size}
}
\value{

}
\description{
Calculate lagged element/delta/ratio
}
