% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_algo.R
\name{make_cumulative_moment}
\alias{make_cumulative_moment}
\title{Cumulative moments for up to 4th order}
\usage{
make_cumulative_moment(order = 4L)
}
\arguments{
\item{order}{order of moments}
}
\value{
a stateful function
}
\description{
Cumulative moments for up to 4th order
}
