% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_order.R
\name{make_cumulative_psquare}
\alias{make_cumulative_psquare}
\title{Online quantile estimation based on P-Square algorithm}
\usage{
make_cumulative_psquare(probs)
}
\arguments{
\item{probs}{}
}
\value{
a stateful online function
}
\description{
Online quantile estimation based on P-Square algorithm
}
