% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_algo.R
\name{make_fame_median}
\alias{make_fame_median}
\title{Fast Algorithm for Median Estimation}
\usage{
make_fame_median(init, b)
}
\arguments{
\item{init}{init value}

\item{b}{minimal initial step}
}
\value{
a stateful function
}
\description{
This is an implementation of Feldman, D., & Shavitt, Y. (2007, May).
An optimal median calculation algorithm for estimating Internet link delays from active measurements.
In 2007 Workshop on End-to-End Monitoring Techniques and Services (pp. 1-7). IEEE.
}
