% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_filter.R
\name{make_filter_rls_linear}
\alias{make_filter_rls_linear}
\title{Create an online RLS filter for MA model}
\usage{
make_filter_rls_linear(width, lambda = 0.99, sigma = 1)
}
\arguments{
\item{width}{width of the filter}

\item{lambda}{forgetting factor}

\item{sigma}{initial factor for covariance matrix}
}
\value{
a stateful online function
}
\description{
Create an online RLS filter for MA model
}
