% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_filter.R
\name{make_filter_rls_poly}
\alias{make_filter_rls_poly}
\title{Create an online RLS filter for polynomial model}
\usage{
make_filter_rls_poly(order, lambda = 0.99, sigma = 1)
}
\arguments{
\item{order}{order of the polynomial}

\item{lambda}{forgetting factor}

\item{sigma}{initial factor for covariance matrix}
}
\value{
a stateful online function
}
\description{
Create an online RLS filter for polynomial model
}
