% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_ma.R
\name{make_hma}
\alias{make_hma}
\title{Hull moving average}
\usage{
make_hma(period, ...)
}
\arguments{
\item{period}{MA period}

\item{...}{not used}
}
\value{
an online function
}
\description{
Hull moving average
}
