% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_ma.R
\name{make_kama}
\alias{make_kama}
\title{Kaufman adaptive moving average}
\usage{
make_kama(period, period_short = 2, period_long = 30, ...)
}
\arguments{
\item{period}{MA period}

\item{period_short}{short period for efficiency ratio}

\item{period_long}{long period for efficiency ratio}

\item{...}{not used}
}
\value{
an online function
}
\description{
Kaufman adaptive moving average
}
