% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_utils.R
\name{make_lag}
\alias{make_lag}
\alias{make_lag_delta}
\alias{make_lag_ratio}
\title{Calculate lagged element/delta/ratio}
\usage{
make_lag(lag)

make_lag_delta(lag)

make_lag_ratio(lag)
}
\arguments{
\item{lag}{lag period}
}
\value{
a stateful online function
}
\description{
Calculate lagged element/delta/ratio
}
