% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_indicator.R
\name{make_macd}
\alias{make_macd}
\title{Moving average convergence/divergence}
\usage{
make_macd(
  period_short = 12L,
  period_long = 26L,
  period_signal = 9L,
  return = c("l", "m")
)
}
\arguments{
\item{period_short}{short EMA period}

\item{period_long}{long EMA period}

\item{period_signal}{signal EMA period}

\item{return}{l for a named list, m for matrix}
}
\value{
a stateful online function
}
\description{
NOTICE: Traditionally, 0.15 and 0.075 are used as 12-day and 26-day EMA factor,
however, this is convention is not used in make_macd. 2 / (period + 1) is used
instead for consistency.
}
