% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_algo.R
\name{make_moving_cov}
\alias{make_moving_cov}
\title{Moving covariance}
\usage{
make_moving_cov(window)
}
\arguments{
\item{window}{moving window size}
}
\value{
a stateful function
}
\description{
Moving covariance
}
