% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_order.R
\name{make_moving_gastwirth}
\alias{make_moving_gastwirth}
\title{Online moving Gastwirth estimator}
\usage{
make_moving_gastwirth(window)
}
\arguments{
\item{window}{moving window size}
}
\value{
a stateful online function
}
\description{
Online moving Gastwirth estimator
}
