% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_stats.R
\name{make_moving_mae}
\alias{make_moving_mae}
\title{Moving mean absolute error}
\usage{
make_moving_mae(window, ...)
}
\arguments{
\item{window}{moving window size}

\item{...}{not used}
}
\value{
a stateful online function
}
\description{
Moving mean absolute error
}
