% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_stats.R
\name{make_moving_mean}
\alias{make_moving_mean}
\alias{make_cumulative_mean}
\alias{make_moving_sd}
\alias{make_cumulative_sd}
\title{Online moving/cumulative mean and sd}
\usage{
make_moving_mean(window, ...)

make_cumulative_mean(...)

make_moving_sd(window, ...)

make_cumulative_sd(...)
}
\arguments{
\item{window}{moving window size}

\item{...}{not used}
}
\value{
a stateful online function
}
\description{
Online moving/cumulative mean and sd
}
