% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_order.R
\name{make_moving_median}
\alias{make_moving_median}
\title{Online moving median}
\usage{
make_moving_median(window)
}
\arguments{
\item{window}{moving window size}
}
\value{
a stateful online function
}
\description{
Online moving median
}
