% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_stats.R
\name{make_moving_moment}
\alias{make_moving_moment}
\alias{make_cumulative_moment}
\alias{make_moving_cov}
\alias{make_cumulative_cov}
\title{Online moving/cumulative statistics}
\usage{
make_moving_moment(window, order, ...)

make_cumulative_moment(order, ...)

make_moving_cov(window, ...)

make_cumulative_cov(...)
}
\arguments{
\item{window}{moving window size}

\item{order}{order of moments}

\item{...}{not used}
}
\value{
a stateful online function
}
\description{
Support up to 4th order central moments.
}
