% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_order.R
\name{make_moving_quantile}
\alias{make_moving_quantile}
\title{Online moving quantile}
\usage{
make_moving_quantile(window, probs)
}
\arguments{
\item{window}{moving window size}

\item{probs}{numeric vector of probabilities with values in [0,1], passed to stats::quantile()}
}
\value{

}
\description{
Only type 3 is supported
}
