% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_order.R
\name{make_moving_sorted}
\alias{make_moving_sorted}
\title{Create a moving sort object.}
\usage{
make_moving_sorted(window)
}
\arguments{
\item{window}{moving window size}
}
\value{
an object of ocls_moving_sort class
}
\description{
Create a moving sort object.
}
