% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_stats.R
\name{make_moving_volatility}
\alias{make_moving_volatility}
\alias{make_cumulative_volatility}
\title{Online moving/cumulative volatility}
\usage{
make_moving_volatility(window, ...)

make_cumulative_volatility(...)
}
\arguments{
\item{window}{moving window size}

\item{...}{not used}
}
\value{
a stateful online function
}
\description{
Online moving/cumulative volatility
}
