% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_ma.R
\name{make_vidya}
\alias{make_vidya}
\alias{vidya}
\title{Variable index dynamic average}
\usage{
make_vidya(period_short = 5, period_long = 20, alpha = 0.2, ...)

vidya(x, period_short = 5L, period_long = 20L, alpha = 0.2)
}
\arguments{
\item{period_short}{short period for efficiency ratio}

\item{period_long}{long period for efficiency ratio}

\item{alpha}{smoothing factor}

\item{...}{not used}

\item{x}{variable to average}
}
\value{
an online function
}
\description{
Variable index dynamic average
}
