% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_volclk.R
\name{make_volclk_order}
\alias{make_volclk_order}
\title{Construct volume clock from order}
\usage{
make_volclk_order(bin, return = c("l", "m"))
}
\arguments{
\item{bin}{volume bin size}

\item{return}{return data type}
}
\value{
a stateful online function
}
\description{
Construct volume clock from order
}
