% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_ma.R
\name{make_wma}
\alias{make_wma}
\title{Weighted moving average}
\usage{
make_wma(period, ...)
}
\arguments{
\item{period}{MA period}

\item{...}{not used}
}
\value{
an online function
}
\description{
Weighted moving average
}
