% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{mass}
\alias{mass}
\title{Mass Index}
\usage{
mass(high, low, period, exp_period = 9L)
}
\arguments{
\item{high}{high}

\item{low}{low}

\item{period}{period}

\item{exp_period}{exponential smoothing period}
}
\value{
numeric vector
}
\description{
Mass Index
}
