% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eod.R
\name{mdaily}
\alias{mdaily}
\title{Query whole-market daily OHLC data}
\usage{
mdaily(trade_date = Sys.Date(), api = TushareApi())
}
\arguments{
\item{trade_date}{trade date to query}

\item{api}{a tsapi object}
}
\value{
a data.table
}
\description{
Query whole-market daily OHLC data
}
