% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{mfi}
\alias{mfi}
\title{Money Flow Index}
\usage{
mfi(price, volume, period)
}
\arguments{
\item{price}{price}

\item{volume}{volume}

\item{period}{period}
}
\value{
numeric vector
}
\description{
Money Flow Index
}
