% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_algo.R
\name{moving_moment}
\alias{moving_moment}
\title{Moving moments for up to 4th order}
\usage{
moving_moment(window, order = 4L)
}
\arguments{
\item{window}{moving window size}

\item{order}{order of mements}
}
\value{
an online function
}
\description{
Moving moments for up to 4th order
}
