% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_algo.R
\name{moving_zscore}
\alias{moving_zscore}
\title{Moving z-score algorithm for anomaly detection}
\usage{
moving_zscore(window, zscore, attenu)
}
\arguments{
\item{window}{moving window size}

\item{zscore}{z-score threshold for signal}

\item{attenu}{attenuation for signal}
}
\value{
an online function
}
\description{
Moving z-score algorithm for anomaly detection
}
