% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_filter.R
\name{new_filter_NLMS}
\alias{new_filter_NLMS}
\title{Create an NLMS (Normalized Least Mean Squares) filter object}
\usage{
new_filter_NLMS(width, mu = 0.1, eps = 1)
}
\arguments{
\item{width}{width of the filter}

\item{mu}{learning factor}

\item{eps}{a positive epsilon to make small norm numeric stable}
}
\value{
an NLMS filter object
}
\description{
Create an NLMS (Normalized Least Mean Squares) filter object
}
