% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_filter.R
\name{new_filter_RLS}
\alias{new_filter_RLS}
\title{Create an RLS (Recursive Least Squares) filter object}
\usage{
new_filter_RLS(width, lambda = 0.99, sigma = 1)
}
\arguments{
\item{width}{width of the filter}

\item{lambda}{forgetting factor}

\item{sigma}{initial factor for covariance matrix}
}
\value{
an RLS filter object
}
\description{
Create an RLS (Recursive Least Squares) filter object
}
