% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_indicator.R
\name{online}
\alias{online}
\alias{make_macd}
\alias{make_ad}
\alias{make_adosc}
\alias{make_ao}
\alias{make_arosc}
\alias{make_atr}
\alias{make_cmo}
\alias{make_dpo}
\alias{make_emv}
\alias{make_mass}
\alias{make_ppo}
\alias{make_rsi}
\alias{make_willr}
\title{Technical indicators, online version}
\usage{
make_macd(
  period_short = 12L,
  period_long = 26L,
  period_signal = 9L,
  return = c("l", "m")
)

make_ad()

make_adosc(period_short, period_long)

make_ao(short_period = 5L, long_period = 34L)

make_arosc(period)

make_atr(period)

make_cmo(period)

make_dpo(period)

make_emv(lot = 100, lag = 1L)

make_mass(period, exp_period = 9L)

make_ppo(short_period, long_period, ma = c("ema", "sma"))

make_rsi(period)

make_willr(period)
}
\arguments{
\item{period_short}{short smoothing period}

\item{period_long}{long smoothing period}

\item{period_signal}{signal period}

\item{return}{return value type}

\item{period}{indicator period}

\item{lot}{volume per lot}

\item{lag}{lag window}

\item{exp_period}{exponential smoothing period}

\item{ma}{moving average method}
}
\value{
a stateful online function
}
\description{
Technical indicators, online version
}
