% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashare_codes.R
\name{parse_ashare_code}
\alias{parse_ashare_code}
\title{Parse A-share codes}
\usage{
parse_ashare_code(
  code,
  type = c("stock", "index", "fund", "conv_bond"),
  SSE = "SH",
  SZSE = "SZ"
)
}
\arguments{
\item{code}{a vector of A-share codes}

\item{type}{type of security}

\item{SSE}{symbol for Shanghai Stock Exchange}

\item{SZSE}{symbol for Shenzhen Stock Exchange}
}
\value{
a list of code and exchange
}
\description{
Parse A-share codes
}
