% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{pnvi}
\alias{pnvi}
\title{Positive/Negative Volume Index}
\usage{
pnvi(close, volume, base_index = 1000, return = c("l", "m"))
}
\arguments{
\item{close}{close}

\item{volume}{volume}

\item{base_index}{base index of the first bar}

\item{return}{return data type, l for list, m for matrix}
}
\value{
a list or matrix
}
\description{
Positive/Negative Volume Index
}
