% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{ppo}
\alias{ppo}
\title{Percentage Price Oscillator}
\usage{
ppo(x, short_period, long_period, ma = c("ema", "sma"))
}
\arguments{
\item{x}{measured variable}

\item{short_period}{short smoothing period}

\item{long_period}{long smoothing period}

\item{ma}{moving average method}
}
\value{
numeric vector
}
\description{
Percentage Price Oscillator
}
