% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tushare_api.R
\name{print.tsapi}
\alias{print.tsapi}
\title{Print Values}
\usage{
\method{print}{tsapi}(x, ...)
}
\arguments{
\item{x}{A tsapi object}

\item{...}{not used}
}
\value{
x, invisibly
}
\description{
Print Values
}
