% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{realtime_js_str_var}
\alias{realtime_js_str_var}
\title{Query function generator}
\usage{
realtime_js_str_var(
  baseurl,
  rand_var,
  code_var,
  encode,
  max_batch,
  var_prefix,
  var_name,
  split,
  cols
)
}
\arguments{
\item{baseurl}{base URL}

\item{rand_var}{random variable name}

\item{code_var}{code variable name}

\item{encode}{content encoding}

\item{max_batch}{max number of codes per batch}

\item{var_prefix}{Javascript assignment variable prefix}

\item{var_name}{variable column name}

\item{split}{Javascript assignment value delimiter}

\item{cols}{column names of the returned data.table}
}
\value{
a function
}
\description{
Query values using HTTP GET. Expected returned values are Javascript assignments
of var_prefix_var_name = "string value"
}
\details{
string values are split and parsed as tabular data.
}
