% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamental.R
\name{report_market}
\alias{report_market}
\title{Query whole-market quarterly reports}
\usage{
report_market(
  type = c("income", "balancesheet", "cashflow", "forecast", "express",
    "fina_indicator", "fina_mainbz"),
  period,
  y,
  q,
  ...,
  api = TushareApi(),
  timeout = 60
)
}
\arguments{
\item{type}{type of report to query}

\item{period}{OPTIONAL, period to query}

\item{y}{OPTIONAL, year to query}

\item{q}{OPTIONAL, quarty to query}

\item{...}{additional arguments passed. Please refer to Tushare manual}

\item{api}{a tsapi object}

\item{timeout}{timeout of the query}
}
\value{
a data.table
}
\description{
Query whole-market quarterly reports
}
