% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamental.R
\name{report_quarter}
\alias{report_quarter}
\title{Query quarterly reports}
\usage{
report_quarter(
  type = c("income", "balancesheet", "cashflow", "forecast", "express",
    "fina_indicator", "fina_mainbz"),
  ts_code,
  ...,
  api = TushareApi(),
  timeout = 60
)
}
\arguments{
\item{type}{type of report to query}

\item{ts_code}{Tushare code}

\item{...}{additional arguments passed. Please refer to Tushare manual}

\item{api}{a tsapi object}

\item{timeout}{timeout of the query}
}
\value{
a data.table
}
\description{
Query quarterly reports
}
