% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample_intraday_ashare}
\alias{resample_intraday_ashare}
\title{Generate resample time intervals by period, A-share version}
\usage{
resample_intraday_ashare(t, period, session = c("day", "morning", "afternoon"))
}
\arguments{
\item{t}{a timestamp, POSIXct and ITime are supported}

\item{period}{resample time period in seconds}

\item{session}{session of trading day}
}
\value{
a vector of POSIXct/ITime depending on t
}
\description{
For A-share intraday only. 9:30 and 1:00 is sampled to reflect orders made
in opening session and lunch break. Extended hour (3:00 - 3:30) is not included.
}
