% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample_ohlc_ashare}
\alias{resample_ohlc_ashare}
\title{Resample A-share OHLC data}
\usage{
resample_ohlc_ashare(dt, period, align_time = FALSE)
}
\arguments{
\item{dt}{OHLC data.table}

\item{period}{resample time period}

\item{align_time}{whether to align time to resampled time intervals}
}
\value{
data.table
}
\description{
Input data dt should have these columns:
Code, Date, Time, Open, High, Low, Close, Vol, Tnvr
}
\details{
Output data contains these columns:
Code, Date, Time, Open, High, Low, Close, Vol, Tnvr, VWAP, Avg

Among which, VWAP stands for volume weighted average price and Avg stands for
average price traded in corresponding bar.
}
