% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{rsi}
\alias{rsi}
\title{Relative Strength Index}
\usage{
rsi(x, period)
}
\arguments{
\item{x}{measured variable}

\item{period}{period}
}
\value{
numeric vector
}
\description{
Relative Strength Index
}
