% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_loop.R
\name{rt_exit_timer}
\alias{rt_exit_timer}
\title{Default exit timer for A-share market}
\usage{
rt_exit_timer(t_now)
}
\arguments{
\item{t_now}{ITime}
}
\value{
TRUE if t_now larger than 15:32:00, FALSE otherwise
}
\description{
Default exit timer for A-share market
}
