% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_loop.R
\name{rt_sleep_timer}
\alias{rt_sleep_timer}
\title{Default sleep timer for A-share market}
\usage{
rt_sleep_timer(t_now)
}
\arguments{
\item{t_now}{ITime}
}
\value{
time in seconds to sleep
}
\description{
Default sleep timer for A-share market
}
