% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdepth.R
\name{select_bookdepth}
\alias{select_bookdepth}
\title{Select bookdepth columns from order book / tick data}
\usage{
select_bookdepth(
  dt,
  book_depth = 5L,
  code_col = "Code",
  date_col = NULL,
  time_col = "Time"
)
}
\arguments{
\item{dt}{order book or tick data}

\item{book_depth}{max book depth}

\item{code_col}{column name of code, defaults to Code}

\item{date_col}{column name of date, defaults to NULL (not selected)}

\item{time_col}{column name of time, defaults to Time}
}
\value{
a data.table
}
\description{
Select bookdepth columns from order book / tick data
}
