% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashare_codes.R
\name{shdjt_ashare_code}
\alias{shdjt_ashare_code}
\title{Query all A-share codes from www.shdjt.com}
\usage{
shdjt_ashare_code()
}
\value{
a data.table
}
\description{
Query all A-share codes from www.shdjt.com
}
