% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashare_codes.R
\name{sina_ashare_code}
\alias{sina_ashare_code}
\title{Normalise A-share codes to Sina format}
\usage{
sina_ashare_code(code, type = c("stock", "index", "fund", "conv_bond"))
}
\arguments{
\item{code}{a vector of A-share codes}

\item{type}{type of security}
}
\value{
normalised codes
}
\description{
Normalise A-share codes to Sina format
}
