% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{sina_realtime}
\alias{sina_realtime}
\title{Query Sina realtime quotes}
\usage{
sina_realtime(api = TushareApi(), sina_code)
}
\arguments{
\item{api}{A tsapi object}

\item{sina_code}{Sina codes to query quotes}
}
\value{
a data.table
}
\description{
Jun's note: It takes about 2 secs to query whole market here in Australia and
time slice of Sina L1 data is about 3 secs.
}
