% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_sina.R
\name{sina_realtime_loop}
\alias{sina_realtime_loop}
\title{Run Sina realtime data loop}
\usage{
sina_realtime_loop(
  codes = default_srt_codes(),
  db = get_srt_db(),
  today = idate_now(),
  api = TushareApi()
)
}
\arguments{
\item{codes}{codes to query}

\item{db}{file path to database}

\item{today}{date of today}

\item{api}{a tsapi object}
}
\value{
this function loops indefinately and does not return
}
\description{
This function simply query newest quotes from Sina and write to an sqlite
database as provided by db argument. Since WAL is used for the database,
please make sure that the file is not on a network location.
}
