% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{sina_realtime_quote}
\alias{sina_realtime_quote}
\title{Query realtime quotes from Sina}
\usage{
sina_realtime_quote(sina_code, api = TushareApi())
}
\arguments{
\item{sina_code}{Sina codes to query}

\item{api}{an tsapi object}
}
\value{
data.table
}
\description{
Query realtime quotes from Sina
}
