% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{sina_realtime_worker}
\alias{sina_realtime_worker}
\title{A simple timed loop to process Sina realtime quote}
\usage{
sina_realtime_worker(
  api = TushareApi(),
  sina_code,
  data_handler,
  data_writer = NULL,
  result_reporter = NULL,
  incremental = TRUE,
  combined = FALSE,
  walltime = 3
)
}
\arguments{
\item{api}{A tsapi object}

\item{sina_code}{A vector of Sina codes}

\item{data_handler}{A data handler function, that accepts a data.table as input}

\item{data_writer}{A data writer function to write recieved Sina quotes}

\item{result_reporter}{A function to report results from data_handler}

\item{incremental}{Whether to pass incremental data to data_handler. If TRUE, only new recieved data is passed, otherwise accumulated history data is passed.}

\item{combined}{Whether to pass all quotes in sina_code combined. If TRUE all quotes are passed to data_handler, otherwise quotes are passed to data_handler grouped by their sina_code.}

\item{walltime}{Timeout for each loop. Sina updates level 1 quotes on a 3 seconds basis.}
}
\value{
a worker function
}
\description{
A simple timed loop to process Sina realtime quote
}
\examples{
codes <- c("sz000001", "sh600000")
worker <- sina_realtime_worker(sina_code = codes, data_handler = example_data_handler, result_reporter = example_result_reporter, combined = TRUE, walltime = 3)
}
