% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdepth.R
\name{spread_bookdepth}
\alias{spread_bookdepth}
\title{Calculate price and volume spread from order book / tick data}
\usage{
spread_bookdepth(dt, book_depth = 5L, reorder = TRUE)
}
\arguments{
\item{dt}{order book or tick data}

\item{book_depth}{max book depth}

\item{reorder}{whether to reorder columns}
}
\value{
a data.table, dt is changed by reference
}
\description{
dt is changed by reference
}
