% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{tbucket_intraday_ashare}
\alias{tbucket_intraday_ashare}
\title{Calculate suitable time buckets/intervals for intraday resampling, A-share version}
\usage{
tbucket_intraday_ashare(t, period, session = c("day", "morning", "afternoon"))
}
\arguments{
\item{t}{a timestamp, POSIXct and ITime are supported}

\item{period}{resample time period in seconds}

\item{session}{session of trading day}
}
\value{
a vector of POSIXct/ITime depending on t
}
\description{
For A-share intraday only. 9:30 and 13:00 is sampled to reflect orders made
in opening session and lunch break. Extended hour (15:00 - 15:30) is not included.
}
