% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{tencent_realtime_moneyflow}
\alias{tencent_realtime_moneyflow}
\title{Query realtime moneyflow data from Tencent}
\usage{
tencent_realtime_moneyflow(tencent_code, api = TushareApi())
}
\arguments{
\item{tencent_code}{Tencent codes to query}

\item{api}{an tsapi object}
}
\value{
data.table
}
\description{
Query realtime moneyflow data from Tencent
}
