% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime.R
\name{tencent_realtime_quote}
\alias{tencent_realtime_quote}
\title{Query realtime quotes from Tencent}
\usage{
tencent_realtime_quote(tencent_code, api = TushareApi())
}
\arguments{
\item{tencent_code}{Tencent codes to query}

\item{api}{as tsapi object}
}
\value{
data.table
}
\description{
Tencent quotes format is not full understood yet. Do not use.
}
