% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_ohlc}
\alias{transform_ohlc}
\title{Transform to OHLC from tick data / order book data}
\usage{
transform_ohlc(dt, vol_acc)
}
\arguments{
\item{dt}{tick data or order book data}

\item{vol_acc}{whether volume data is accumulated, auto detected if not supplied}
}
\value{
a data.table
}
\description{
Input data dt should have these columns:
Code, Date, Time, Price, Volume (Vol), Turnover (Tnvr)
}
\details{
Output data contains these columns:
Code, Date, Time, Open, High, Low, Close, Vol, Tnvr, VWAP, Avg

Note: this function sets key to dt on Code, Date, Time
}
