% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_price}
\alias{transform_price}
\title{Transform to price distribution from tick data / order book data}
\usage{
transform_price(dt, vol_acc)
}
\arguments{
\item{dt}{tick data or order book data}

\item{vol_acc}{whether volume data is accumulated, auto detected if not supplied}
}
\value{
a data.table
}
\description{
Input data dt should have these columns:
Code, Date, Time, Price, Volume (Vol), Turnover (Tnvr)
}
\details{
Output data contains these columns:
Code, Date, Price, Vol, Tnvr
}
