% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{trix}
\alias{trix}
\title{Trix Momentum Indicator}
\usage{
trix(x, period)
}
\arguments{
\item{x}{measured variable}

\item{period}{period}
}
\value{
numeric vector
}
\description{
Trix Momentum Indicator
}
