% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{trsmp_intraday}
\alias{trsmp_intraday}
\title{Calculate suitable time intervals for intraday resampling}
\usage{
trsmp_intraday(t, period, session)
}
\arguments{
\item{t}{a timestamp, POSIXct and ITime are supported}

\item{period}{resample time period in seconds}

\item{session}{session definition}
}
\value{
a vector of POSIXct/ITime depending on t
}
\description{
Trading hours are defined in session, a list of integer vectors, each defining
a trading session by 6 integers: session start hour, minute, second, session end
hour, minute, second.
}
