% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_def.R
\name{tsapi_usage}
\alias{tsapi_usage}
\title{Return usage for a Tushare API function}
\usage{
tsapi_usage(func, what = c("field", "arg"))
}
\arguments{
\item{func}{function name}

\item{what}{field: show returned fields, arg: show API arguments}
}
\value{
data.table
}
\description{
Return usage for a Tushare API function
}
