% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{tsec_intraday_ashare}
\alias{tsec_intraday_ashare}
\title{Convert A-share intraday timestamps to seconds traded}
\usage{
tsec_intraday_ashare(t)
}
\arguments{
\item{t}{a vector of timestamps}
}
\value{
a vector of integer, counted seconds of t
}
\description{
9:30 is converted to 0th second, 13:00 is converted to 7201st second
}
