% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_ts.R
\name{tushare_realtime_ping}
\alias{tushare_realtime_ping}
\title{Ping Tushare websocket}
\usage{
tushare_realtime_ping(ws)
}
\arguments{
\item{ws}{a WebSocket}
}
\value{
TRUE
}
\description{
Ping Tushare websocket
}
