% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_ts.R
\name{tushare_realtime_start}
\alias{tushare_realtime_start}
\alias{tushare_realtime_stop}
\alias{tushare_realtime_ping}
\title{Start/Stop/Ping Tushare websocket}
\usage{
tushare_realtime_start(ws, timeout = 10)

tushare_realtime_stop(ws)

tushare_realtime_ping(ws)
}
\arguments{
\item{ws}{a Websocket created by tushare_realtime_websocket()}

\item{timeout}{timeout}
}
\value{
TRUE
}
\description{
Start/Stop/Ping Tushare websocket
}
