% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_ts.R
\name{tushare_realtime_tick_parser}
\alias{tushare_realtime_tick_parser}
\title{Generate a record parser for Tushare realtime tick data}
\usage{
tushare_realtime_tick_parser(api = TushareApi())
}
\arguments{
\item{api}{a tsapi object}
}
\value{
a parser function
}
\description{
Generate a record parser for Tushare realtime tick data
}
