% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_ts.R
\name{tushare_realtime_websocket}
\alias{tushare_realtime_websocket}
\title{Create a Tushare realtime websocket}
\usage{
tushare_realtime_websocket(topic, code, callback, api = TushareApi())
}
\arguments{
\item{topic}{realtime topic to subscribe}

\item{code}{code to subscribe}

\item{callback}{callback function to process data}

\item{api}{an tsapi object}
}
\value{
a WebSocket
}
\description{
Create a Tushare realtime websocket
}
