% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tushare_api.R
\name{TushareApi}
\alias{TushareApi}
\title{Get a Tushare API object.}
\usage{
TushareApi(
  api_token = GetToken(),
  time_mode = c("POSIXct", "char"),
  date_mode = c("Date", "POSIXct", "char"),
  logi_mode = c("logical", "char"),
  tz = "Asia/Shanghai"
)
}
\arguments{
\item{api_token}{API token.}

\item{time_mode}{data type for time objects}

\item{date_mode}{data type for date objects}

\item{logi_mode}{data type for logical objects}

\item{tz}{Default timezone of POSIXct data}
}
\value{
a tsapi object.
}
\description{
Get a Tushare API object.
}
