% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tushare_api.R
\name{TusRequest}
\alias{TusRequest}
\title{Make raw request to Tushare Pro API}
\usage{
TusRequest(
  api_name,
  ...,
  fields = c(""),
  token = GetToken(),
  timeout = 10,
  attempt = 3L
)
}
\arguments{
\item{api_name}{name of API function, please refer to online document for more information.}

\item{...}{passed to API function.}

\item{fields}{data fields to request}

\item{token}{API token.}

\item{timeout}{timeout in seconds for httr request.}

\item{attempt}{attempts if timeout}
}
\value{
data.frame/data.table
}
\description{
Make raw request to Tushare Pro API
}
