% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_clock.R
\name{vclock_value}
\alias{vclock_value}
\title{Construct volume clock by total value}
\usage{
vclock_value(dt, val, p_col = "Price", v_col = "Volume")
}
\arguments{
\item{dt}{a data.table/data.frame}

\item{val}{total value per tick}

\item{p_col}{price colume}

\item{v_col}{volume colume}
}
\value{
a data.table of average price, high and low
}
\description{
Construct volume clock by total value
}
