% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{willr}
\alias{willr}
\title{Williams %R}
\usage{
willr(high, low, close, period)
}
\arguments{
\item{high}{high}

\item{low}{low}

\item{close}{close}

\item{period}{period}
}
\value{
numeric vector
}
\description{
Williams %R
}
