% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eod.R
\name{adjust_ohlc}
\alias{adjust_ohlc}
\title{Perform dividen adjustment on OHLC data}
\usage{
adjust_ohlc(api, ohlc, adjust = c("forward", "qfq", "backward", "hfq"))
}
\arguments{
\item{api}{a tsapi object}

\item{ohlc}{a data.table of OHLC data}

\item{adjust}{Adjustment method}
}
\value{
a data.table
}
\description{
Perform dividen adjustment on OHLC data
}
