% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dapdv.R
\name{dapdv}
\alias{dapdv}
\title{Estimate Target Price}
\usage{
dapdv(
  api,
  ts_code,
  div_rate,
  discount_rate,
  N = 3,
  start_date = "",
  intraday_freq = 15,
  intraday_bar = 4000,
  arb_end = ""
)
}
\arguments{
\item{api}{a tsapi object}

\item{ts_code}{Tushare code}

\item{div_rate}{estimated dividen rate}

\item{discount_rate}{discount rate}

\item{N}{year of stable growth}

\item{start_date}{start date of historic growth}

\item{intraday_freq}{intraday data frequency}

\item{intraday_bar}{bars of intraday data used}

\item{arb_end}{an arbitrary end date for testing only}
}
\value{
a data.table
}
\description{
Discounted Assets on Probability Distribution is an experimental valuation
method to estimate the value of an investment based on its future assets.
DAPD is determined by estimated growth from historic ROE and market valuation
preference from distribution of recent intraday PB values.
}
\details{
Returned value is a data.table of current prices (discounted) and target prices
in N years.

Data is provided by Tushare Pro API.
}
\examples{
\dontrun{
    api <- TushareApi("YOUR TUSHARE API TOKEN")
    #Valuate 601117.SH base on 3 yr growth. Dividen rate is simply taken from
    #latest data and discount rate is assumed 10\%
    dapdv(api, "601117.SH", 0.0191 * 11.08, 0.1, 3, start_date = "2010-01-01")
}
}
