% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{indicators}
\alias{indicators}
\title{Technical indicators}
\description{
Various technical indicators and their online versions are provided:
}
\details{
macd: 	Moving Average Convergence/Divergence

ad: Accumulation/Distribution Line

adosc: Accumulation/Distribution Oscillator

ao: Awesome Oscillator

arosc: Aroon Oscillator

atr: Average True Range

cci: Commodity Channel Index

cmo: Chande Momentum Oscillator

dpo: Detrended Price Oscillator

emv: Ease of Movement

mass: Mass Index

mfi: Money Flow Index

obv: On Balance Volume

pnvi: Positive/Negative Volume Index

ppo: Percentage Price Oscillator

rsi: Relative Strength Index

trix: Trix Momentum Indicator

willr: Williams %R
}
