% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{insert_to}
\alias{insert_to}
\title{Insert data to an SQL connection}
\usage{
insert_to(
  con,
  tbl,
  dt,
  conflict = c("default", "replace", "ignore", "do_update", "do_nothing")
)
}
\arguments{
\item{con}{an SQL connection}

\item{tbl}{name of table to insert to}

\item{dt}{data to insert}

\item{conflict}{how to deal with conflict}
}
\value{
number of rows inserted
}
\description{
Insert data to an SQL connection
}
