% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_order.R
\name{make_cumulative_quantile}
\alias{make_cumulative_quantile}
\title{Online quantile estimation based on KLL algorithm}
\usage{
make_cumulative_quantile(probs, k = 128L)
}
\value{
a stateful online function
}
\description{
Online quantile estimation based on KLL algorithm
}
