% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_ma.R
\name{make_ema}
\alias{make_ema}
\alias{make_dema}
\alias{make_tema}
\alias{make_zlema}
\title{Online exponential moving average}
\usage{
make_ema(period, ...)

make_dema(period, ...)

make_tema(period, ...)

make_zlema(period, ...)
}
\arguments{
\item{period}{MA period}

\item{...}{not used}
}
\value{
an online function
}
\description{
Avalable variances are:
make_ema: normal EMA
make_dema: double EMA
make_tema: triple EMA
make_zlema: zero lag EMA
}
