% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_ma.R
\name{make_hma}
\alias{make_hma}
\alias{hma}
\title{Hull moving average}
\usage{
make_hma(period, ...)

hma(x, period)
}
\arguments{
\item{period}{MA period}

\item{...}{not used}

\item{x}{variable to average}
}
\value{
an online function
}
\description{
Hull moving average
}
