% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_ma.R
\name{make_vwma}
\alias{make_vwma}
\title{Volume weighted moving average}
\usage{
make_vwma(period, ...)
}
\arguments{
\item{period}{MA period}

\item{...}{not used}
}
\value{
an online function
}
\description{
Volume weighted moving average
}
