% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eod.R
\name{market_eod}
\alias{market_eod}
\title{Query whole-market end-of-day data by date}
\usage{
market_eod(
  func,
  date = Sys.Date(),
  freq = c("60", "15", "5", "1"),
  ...,
  api = TushareApi()
)
}
\arguments{
\item{func}{Tushare function to call. Use intraday for intraday data.}

\item{date}{date to query}

\item{freq}{only used when func == "intraday", intraday data frequency}

\item{...}{other arguments passed to the query}

\item{api}{a tsapi object}
}
\value{
a data.table object, keyed by ts_code (and trade_time if the query is intraday)
}
\description{
Query whole-market end-of-day data by date
}
