% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eod.R
\name{mdaily}
\alias{mdaily}
\title{Query whole-market daily OHLC data}
\usage{
mdaily(api, trade_date = Sys.Date())
}
\arguments{
\item{api}{a tsapi object}

\item{trade_date}{trade date to query}
}
\value{
a data.table
}
\description{
Query whole-market daily OHLC data
}
