% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{select_from}
\alias{select_from}
\title{Select from an SQL connection}
\usage{
select_from(con, tbl, what = "*", ...)
}
\arguments{
\item{con}{an SQL connection}

\item{tbl}{name of table to query}

\item{what}{what variables/columns to query}

\item{...}{named arguments, parsed as WHERE AND clause}
}
\value{
data.table
}
\description{
Select from an SQL connection
}
