% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tushare_api.R
\name{SetToken}
\alias{SetToken}
\alias{GetToken}
\title{Set/Get Tushare Pro API token}
\usage{
SetToken(token, save = FALSE, token_file = "~/tk.csv")

GetToken(token_file = "~/tk.csv")
}
\arguments{
\item{token}{a character vector}

\item{save}{save set token to token_file}

\item{token_file}{path to token_file}
}
\value{
token itself, invisibly.
}
\description{
Set/Get Tushare Pro API token
}
