% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_sina.R
\name{sina_realtime_loader}
\alias{sina_realtime_loader}
\title{Generate a Sina realtime data loader function}
\usage{
sina_realtime_loader(db = get_srt_db(), today = Sys.Date(), api = TushareApi())
}
\arguments{
\item{db}{path to database file}

\item{today}{date of today}

\item{api}{a tsapi object}
}
\value{
a load function
}
\description{
The generated function reads today's incremental data loaded to db. It also
takes an optional argument dt_bind, if supplied with a data.table/data.frame,
the load rbind old dt_bind data to newly loaded data, constructing all-day
data
}
