% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_ts.R
\name{tushare_realtime_start}
\alias{tushare_realtime_start}
\title{Start Tushare websocket}
\usage{
tushare_realtime_start(ws, timeout = 10)
}
\arguments{
\item{ws}{a Websocket created by tushare_realtime_websocket()}

\item{timeout}{timeout}
}
\value{
TRUE
}
\description{
Start Tushare websocket
}
